requirements-ftp.txt
# FTP Server Dependencies
pyftpdlib>=1.5.7
python-dotenv>=1.0.0
